pyproject.toml
[project]
name = "my_agent"
version = "0.1.0"
description = "A LangChain and LangGraph based AI agent with RAG, SQL, and web search capabilities"
readme = "README.md"
requires-python = ">=3.12"
license = {text = "MIT"}
authors = [
    {name = "trinstar", email = "tannyecho@gmail.com"}
]
keywords = ["langchain", "langgraph", "ai", "agent", "rag", "llm"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    # Core runtime deps from requirements.txt
    "dashscope==1.24.9",
    "faiss-gpu-cu12==1.12.0",
    "fastapi==0.120.4",
    "flightradar24==0.3.1",
    "fqdn==1.5.1",
    "hf_transfer==0.1.9",
    "isoduration==20.11.0",
    "jupyter==1.1.1",
    "langchain==1.1.0",
    "langchain-community==0.4.1",
    "langchain-deepseek==1.0.1",
    "langchain-huggingface==1.1.0",
    "langchain-mcp-adapters==0.1.14",
    "langchain-ollama==1.0.0",
    "langgraph-api==0.5.27",
    "langgraph-cli==0.4.7",
    "onnxruntime==1.19.2",
    "rank-bm25==0.2.2",
    "rfc3987-syntax==1.1.0",
    "rich==14.2.0",
    "sentence-transformers==5.1.2",
    "tinycss2==1.4.0",
    "torchaudio==2.5.1+cu121",
    "torchvision==0.20.1+cu121",
    "unstructured-client==0.42.3",
    "uri-template==1.3.0",
    "webcolors==25.10.0",
    "yagmail==0.15.293",
]

[project.optional-dependencies]
dev = [
    # Dev/test/analysis tools
    "pipdeptree==2.29.0",
    "pytest==9.0.1",
    "ipython>=9.7.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["my_agent*"]

[tool.setuptools.package-data]
my_agent = ["utils/rag/config.yaml"]

[tool.black]
line-length = 100
target-version = ["py312"]
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
  | __pycache__
)/
'''

[tool.ruff]
line-length = 100
target-version = "py312"
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "C",  # flake8-comprehensions
    "B",  # flake8-bugbear
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q"
testpaths = ["tests"]
pythonpath = ["."]
asyncio_mode = "auto"

[tool.coverage.run]
source = ["my_agent"]
omit = ["*/tests/*", "*/__pycache__/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
]
